wip: dynamic gas setting
